Use Maven repository within workspace
For whatever reason we're using custom Maven repository, it might be
that we don't want two builds in parallel to interfere with each other.
Using the same repository (within $HOME) then doesn't make much sense.
So if the motive was to isolate the build, local repository should be
within the Jenkins $WORKSPACE.