ENTESB-18360: Upgrade Pulsar to 2.9.1
This upgrades Pulsar to 2.9.1 (from 2.2.1) in order to use the
`apachepulsar/pulsar:2.9.1` (from the default 2.2.0) container in the
tests. Upgrading just the container used, most likely because of the
large version difference, lead to client not being able to connect with
error:

```
org.apache.pulsar.client.api.PulsarClientException$BrokerMetadataException: Policies not found for public/default namespace
```

This doesn't back port a number of changes made upstream in order to
keep the change to a minimum.

This is possibly a breaking change, due to a change in the version of
the client and a change in the Java API used to configure the client as
demonstrated by the test failures and a change in the
`AutoConfiguration`.